requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.16
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-js-asset==1.2.2
django-phonenumber-field==5.2.0
gunicorn==20.1.0
phonenumbers==8.12.29
Pillow==9.0.1
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.2
requests==2.28.2
whitenoise==5.3.0
